requirements.txt
playwright
google-generativeai
supabase
streamlit
python-dotenv
pandas
requests
